# Conflicts:
#	Admin/target/classes/com/example/admin/Service/Impl/CategoryServiceImpl.class
#	Admin/target/classes/static/css/app.css

+++ Admin/src/main/java/com/example/admin/Domain/AmountData.java
@@ -7,5 +7,5 @@ import lombok.Setter;
 @Setter
 public class AmountData {
     private String month;
-    private double amount;
+    private Long amount;
 }
