+++ Admin/src/main/java/com/example/admin/Entity/VoucherEntity.java
@@ -15,15 +15,26 @@ public class VoucherEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+
     private String title;
+
     private String icon;
+
     private String code;
-    private double percentage;
+
+    private Long percentage;
+
     private int usageLimit;
+
     private LocalDate startDate;
+
     private LocalDate endDate;
+
     private boolean conditionsPaymentOnline;
+
     private double conditionPrice;
+
     private boolean expired;
+
     private VoucherType type;
 }
